Remove PatriciaTriePath class altogether
